Separate tests using jmockit from jersey core
