+++ tests/jmockit/src/main/java/org/glassfish/jersey/tests/jmockit/server/innerstatic/InnerStaticClass.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) 2014, 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+
+package org.glassfish.jersey.tests.jmockit.server.innerstatic;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+
+/**
+ * @author Paul Sandoz
+ */
+public class InnerStaticClass {
+
+    @Path("/")
+    public static class PublicClass {
+
+        @GET
+        public String getMe() {
+            return "ME";
+        }
+    }
+
+    @Path("/")
+    private static class PrivateClass {
+
+        @GET
+        public String getMe() {
+            return "ME";
+        }
+    }
+
+    @Path("/")
+    protected static class ProtectedClass {
+
+        @GET
+        public String getMe() {
+            return "ME";
+        }
+    }
+
+    @Path("/")
+    /* package */ static class PackageClass {
+
+        @GET
+        public String getMe() {
+            return "ME";
+        }
+    }
+}
