+++ tests/pom.xml
@@ -45,6 +45,7 @@
         <module>e2e-server</module>
         <module>e2e-testng</module>
         <module>integration</module>
+        <module>jmockit</module>
         <module>mem-leaks</module>
         <module>osgi</module>
         <module>stress</module>
